requirements.txt
certifi==2018.11.29
chardet==3.0.4
dj-database-url==0.5.0
Django==2.1.5
django-heroku==0.3.1
future==0.17.1
gunicorn==19.9.0
idna==2.8
Janome==0.3.7
line-bot-sdk==1.8.0
psycopg2==2.7.7
pytz==2018.9
requests==2.21.0
urllib3==1.24.1
whitenoise==4.1.2
